refactor: move docker logic to new docker package
