+++ pkg/actions/status.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/eclipse/codewind-installer/pkg/apiroutes"
 	"github.com/eclipse/codewind-installer/pkg/connections"
-	"github.com/eclipse/codewind-installer/pkg/utils"
+	"github.com/eclipse/codewind-installer/pkg/docker"
 	"github.com/urfave/cli"
 )
 
@@ -86,13 +86,13 @@ func StatusCommandRemoteConnection(c *cli.Context) {
 
 // StatusCommandLocalConnection : Output local connection details
 func StatusCommandLocalConnection(c *cli.Context) {
-	dockerClient, dockerErr := utils.NewDockerClient()
+	dockerClient, dockerErr := docker.NewDockerClient()
 	if dockerErr != nil {
 		HandleDockerError(dockerErr)
 		os.Exit(1)
 	}
 
-	containersAreRunning, err := utils.CheckContainerStatus(dockerClient)
+	containersAreRunning, err := docker.CheckContainerStatus(dockerClient)
 	if err != nil {
 		HandleDockerError(err)
 		os.Exit(1)
@@ -100,19 +100,19 @@ func StatusCommandLocalConnection(c *cli.Context) {
 
 	if containersAreRunning {
 		// Started
-		hostname, port, err := utils.GetPFEHostAndPort(dockerClient)
+		hostname, port, err := docker.GetPFEHostAndPort(dockerClient)
 		if err != nil {
 			HandleDockerError(err)
 			os.Exit(1)
 		}
 		if printAsJSON {
-			imageTagArr, err := utils.GetImageTags(dockerClient)
+			imageTagArr, err := docker.GetImageTags(dockerClient)
 			if err != nil {
 				fmt.Println(err.Error())
 				os.Exit(1)
 			}
 
-			containerTagArr, err := utils.GetContainerTags(dockerClient)
+			containerTagArr, err := docker.GetContainerTags(dockerClient)
 			if err != nil {
 				fmt.Println(err.Error())
 				os.Exit(1)
@@ -140,7 +140,7 @@ func StatusCommandLocalConnection(c *cli.Context) {
 		os.Exit(0)
 	}
 
-	imagesAreInstalled, err := utils.CheckImageStatus(dockerClient)
+	imagesAreInstalled, err := docker.CheckImageStatus(dockerClient)
 	if err != nil {
 		HandleDockerError(err)
 		os.Exit(1)
@@ -150,7 +150,7 @@ func StatusCommandLocalConnection(c *cli.Context) {
 		// Installed but not started
 		if printAsJSON {
 
-			imageTagArr, err := utils.GetImageTags(dockerClient)
+			imageTagArr, err := docker.GetImageTags(dockerClient)
 			if err != nil {
 				fmt.Println(err.Error())
 				os.Exit(1)
