+++ pkg/actions/stop.go
@@ -15,7 +15,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/eclipse/codewind-installer/pkg/utils"
+	"github.com/eclipse/codewind-installer/pkg/docker"
 	"github.com/urfave/cli"
 )
 
@@ -23,7 +23,7 @@ import (
 func StopCommand(c *cli.Context, dockerComposeFile string) {
 	tag := c.String("tag")
 	fmt.Println("Only stopping Codewind containers. To stop project containers, please use 'stop-all'")
-	err := utils.DockerComposeStop(tag, dockerComposeFile)
+	err := docker.DockerComposeStop(tag, dockerComposeFile)
 	if err != nil {
 		HandleDockerError(err)
 		os.Exit(1)
