+++ pkg/utils/filesystem_test.go
@@ -116,33 +116,6 @@ func TestDirIsEmpty(t *testing.T) {
 	})
 }
 
-func TestWriteToComposeFile(t *testing.T) {
-	t.Run("docker compose should be written to the filepath", func(t *testing.T) {
-		testFile := "TestFile.yaml"
-		os.Create(testFile)
-		_ = WriteToComposeFile("TestFile.yaml", false)
-
-		pathExists := PathExists(testFile)
-		assert.True(t, pathExists)
-		os.Remove(testFile)
-	})
-
-	t.Run("docker compose should be written to the filepath", func(t *testing.T) {
-		testFile := "TestFile.yaml"
-		os.Create(testFile)
-		composeWritten := WriteToComposeFile("TestFile.yaml", false)
-
-		pathExists := PathExists(testFile)
-		assert.True(t, pathExists)
-		assert.True(t, composeWritten)
-		os.Remove(testFile)
-	})
-	t.Run("empty path returns nil", func(t *testing.T) {
-		composeWritten := WriteToComposeFile("", false)
-		assert.False(t, composeWritten)
-	})
-}
-
 func TestReplaceInFiles(t *testing.T) {
 	t.Run("replaces placeholder in file", func(t *testing.T) {
 		testFile, removeFile := CreateTempTestFile(t, "[PROJ_NAME_PLACEHOLDER] test")
